Create Scala 2 veneer for probe/Probe.scala
